@feature/86epe3476 #HCP Questionnaire refactoring to blocks
