+++ apps/simulator/src/components/blocks/CaseStudy/CaseStudyQuestions/index.ts
@@ -1,10 +1,11 @@
 export * from './DDCQuestionire'
 export * from './DDTQuestionaire'
-export * from './HCPQuestionaire'
+export * from './HCP/HCPBlock'
 export * from './MRSNQuestionaire'
 export * from './MCQCSQuestionnaire'
 export * from './MCQCSQuestionnaire'
 export * from './DNDQuestionaire'
 export * from './MCQTable/MCQAnswerGroupTable'
 export * from './DNDBowtieQuestionaire'
+
